Some valgrind and compilation (clang and Wall) cleanup      related to https://lists.quagga.net/pipermail/quagga-dev/2013-April/010455.html

+++ quagga/ospfd/ospf_zebra.c
@@ -1019,7 +1019,7 @@ ospf_distribute_list_update (struct ospf *ospf, uintptr_t type)
   /* Set timer. */
   ospf->t_distribute_update =
     thread_add_timer (master, ospf_distribute_list_update_timer,
-                      (void *) type, OSPF_DISTRIBUTE_UPDATE_DELAY);
+                      (void *) ((long) type), OSPF_DISTRIBUTE_UPDATE_DELAY);
 }
 
 /* If access-list is updated, apply some check. */
